Reconcile ClusterAdmissionPolicy status in a separate controller
By reconciling ClusterAdmissionPolicies in its own loop we focus on
observing the current status of the ConfigMap created for the current
PolicyServer, and update `ClusterAdmissionPolicies` statuses based on
that.

This way, a user can be certain that after a policy has been
reconfigured (e.g. transitioned from `monitor` to `protect` mode -- or
settings have been changed), they can wait for the
`PolicyServerConfigurationUpToDate` condition, that will signal that
the configuration is up to date on the PolicyServer after any change
has been made to a `ClusterAdmissionPolicy`. After the user has waited
for this condition, they can wait for the rollout of the policy server
deployment to ensure that the policy in the right mode (or with the
right policy settings) is finally active.