Add `scheduled` status on policies
This status signals when a policy is bound to a policy server,
regardless of the existance or status of the policy server itself.

Then the policy will transition to either `active` or `failed`. If the
policy server does not exist, it will be kept in `scheduled` status
until the matching policy server is created.